# Add "CreateReadOnlySpanFromNullTerminated" on NET6-specific path