@Havret 2025-07-10
This can be closed as https://github.com/confluentinc/confluent-kafka-dotnet/pull/2466 was already merged.